refactor: 패키지 구조 변경
- ui 안의 main에 있던 파일들의 구조를 변경
- ui/main 내부의 패키지를 외부로 옮김
