+++ app/src/androidTest/java/woowacourse/movie/ui/main/MainActivityTest.kt
@@ -26,6 +26,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
+import woowacourse.movie.main.MainActivity
 import woowacourse.movie.ui.adv.AdvDetailActivity
 import woowacourse.movie.ui.reservation.MovieDetailActivity
