+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
             android:name=".ui.reservation.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
