+++ app/src/main/java/woowacourse/movie/model/MovieState.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.model
 import android.os.Parcelable
 import androidx.annotation.DrawableRes
 import kotlinx.parcelize.Parcelize
-import woowacourse.movie.ui.main.itemModel.MovieItemModel
+import woowacourse.movie.ui.itemModel.MovieItemModel
 import java.time.LocalDate
 
 @Parcelize
