+++ app/src/main/java/woowacourse/movie/model/TicketsState.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.movie.ui.main.itemModel.TicketsItemModel
+import woowacourse.movie.ui.itemModel.TicketsItemModel
 import java.time.LocalDateTime
 
 @Parcelize
