+++ app/src/main/java/woowacourse/movie/ui/adapter/ReservationListAdapter.kt
@@ -1,12 +1,12 @@
-package woowacourse.movie.ui.main.adapter
+package woowacourse.movie.ui.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.ui.main.itemModel.ItemModel
-import woowacourse.movie.ui.main.viewHolder.ItemViewHolder
-import woowacourse.movie.ui.main.viewHolder.TicketsViewHolder
+import woowacourse.movie.ui.itemModel.ItemModel
+import woowacourse.movie.ui.viewHolder.ItemViewHolder
+import woowacourse.movie.ui.viewHolder.TicketsViewHolder
 
 class ReservationListAdapter(
     reservations: List<ItemModel>
