+++ app/src/main/java/woowacourse/movie/ui/confirm/AlarmReceiver.kt
@@ -4,14 +4,14 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import woowacourse.movie.PreferenceUtil
+import woowacourse.movie.util.PreferenceUtil
 import woowacourse.movie.R
 import woowacourse.movie.model.TicketsState
 import woowacourse.movie.ui.fragment.setting.SettingFragment.Companion.NOTIFICATIONS
 import woowacourse.movie.util.getParcelableCompat
 import woowacourse.movie.util.sendNotification
 
-class AlarmReceiver() : BroadcastReceiver() {
+class AlarmReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent?) {
         val tickets = intent?.extras?.getParcelableCompat<TicketsState>("a")
