+++ app/src/main/java/woowacourse/movie/ui/viewHolder/ItemViewHolder.kt
@@ -1,8 +1,8 @@
-package woowacourse.movie.ui.main.viewHolder
+package woowacourse.movie.ui.viewHolder
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.movie.ui.main.itemModel.ItemModel
+import woowacourse.movie.ui.itemModel.ItemModel
 
 abstract class ItemViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     abstract fun bind(itemModel: ItemModel)
